Everything I've done so far- see description
For directions, see the comment in client/index.html
I'm not sure if you need to install node.js on your computer, but it
can't hurt. You probably don't need to install libraries.
I've made it so that each player is a character which moves by ASDW on a
capture the flag board, where one half is red and the other half is
blue.
There is also a chat box which can be used to debug things if your
message starts with "/"